requirements.txt
requests~=2.31.0
numpy~=1.26.3
pandas~=2.1.4
tqdm~=4.65.0
pillow~=10.0.1
scikit-learn~=1.3.0
matplotlib~=3.8.0
seaborn~=0.12.2
torch